+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Firebase: add `ignorePrimaryKey` option
+
 ## 0.2.2 (2017-5-15)
 
 - Firebase: include root `rules` property